Move Home functionalies from index to home file
